Refactor User data class
Since we need to differentiate between Customer and Employee, I have
made the User class abstract and let the subclasses extend that data
entity.

The outcome is that both Customer and Employee have a different table.
We are duplicating data, however. But in our case this is good enough
for now.