Update data layer
Modified and added a few new entities. FlightSeatEntity is a special
case, since it is actually used as a coupling table between SeatEntity,
PassengerEntity and FlightEntity.

This way, the seats added to an aircraft will be able to keep track when
and which flight it was occupied etc.

Also added a few new Repositories.