Add MapStruct dependency
This makes it easier to map certain classes, I noticed that we were
unpacking lots of dto's and then calling the service with those parameters.
That is why I added a 'struct' class (naming subject to change), to hold
the data. It is in its own layer, that's why we can use it.

Lombok is also moved to the annotationProcessorPath in pom.xml.